Switch from pyo3 to maturin, and divert versioning with startin
The build with maturin is way cleaner and simpler, and all platforms
can be build with GitHub Actions easily.

The bindings should have their own version, so that they can be updated
even when the Rust code of startin is not. 0.6.0 is very arbitrary, I
just picked a number.

This removes the Travis CI and Appveyor yml files too.